Create order management helper class
